refactored users-messages to blueprint
