+++ Unit-02/01-blueprints/manage.py
@@ -1,4 +1,4 @@
-from app import app,db 
+from project import app,db 
 from flask_script import Manager
 from flask_migrate import Migrate, MigrateCommand
