+++ Unit-02/01-blueprints/project/messages/models.py
@@ -1,0 +1,13 @@
+from project import db
+
+class Message(db.Model):
+
+	__tablename__ = 'messages'
+
+	id = db.Column(db.Integer, primary_key=True)
+	text = db.Column(db.Text)
+	user_id = db.Column(db.Integer, db.ForeignKey('users.id'))
+
+	def __init__(self, text, user_id):
+		self.text = text
+		self.user_id = user_id
